+++ Robust.Shared/CVars.cs
@@ -6,6 +6,7 @@ using Robust.Shared.Audio;
 using Robust.Shared.Configuration;
 using Robust.Shared.GameObjects;
 using Robust.Shared.Log;
+using Robust.Shared.Maths;
 using Robust.Shared.Network;
 using Robust.Shared.Physics;
 
@@ -761,6 +762,9 @@ namespace Robust.Shared
         public static readonly CVarDef<double> RenderSpriteDirectionBias =
             CVarDef.Create("render.sprite_direction_bias", -0.05, CVar.ARCHIVE | CVar.CLIENTONLY);
 
+        public static readonly CVarDef<string> RenderFOVColor =
+            CVarDef.Create("render.fov_color", Color.Black.ToHex(), CVar.ARCHIVE | CVar.CLIENTONLY);
+
         /*
          * DISPLAY
          */
@@ -1226,6 +1230,15 @@ namespace Robust.Shared
         /// </remarks>
         public static readonly CVarDef<bool> ResPrototypeReloadWatch =
             CVarDef.Create("res.prototype_reload_watch", true, CVar.CLIENTONLY);
+
+        /// <summary>
+        /// If true, do a warning check at startup for probably-erroneous file extensions like <c>.yaml</c> in resources.
+        /// </summary>
+        /// <remarks>
+        /// This check is always skipped on <c>FULL_RELEASE</c>.
+        /// </remarks>
+        public static readonly CVarDef<bool> ResCheckBadFileExtensions =
+            CVarDef.Create("res.check_bad_file_extensions", true);
 
         /*
          * DEBUG