+++ RELEASE-NOTES.md
@@ -41,7 +41,7 @@ END TEMPLATE-->
 
 ### New features
 
-*None yet*
+* The FOV color is now configurable via the "render.fov_color" cvar
 
 ### Bugfixes
 
@@ -51,6 +51,7 @@ END TEMPLATE-->
 
 * SDL2 backend now handles quit events (⌘+Q on macOS).
 * SDL2 backend now logs video driver backend used on initialization.
+* The engine will now warn on startup if `*.yaml` files are found in resources, as this most likely indicates an accident.
 * Added entity, occluder and shadow-casting light counts to the clyde debug panel.
 
 ### Internal